requirements.txt
numpy==1.23.3
PyYAML==6.0.1
scikit_learn==1.2.0
torch==1.13.0
torch_geometric==2.3.0
tqdm==4.64.1

# only for visualisation purposes
matplotlib==3.5.2
networkx==2.8.4